refactor(resourceset): split render.go into per-concern files
Same per-concern split pattern as #253/#255/#256. The 523 LOC
render.go mixed five concerns; this commit extracts three clusters
out alongside the existing permute.go sibling.

New layout:

  pkg/resourceset/
    render.go    (Render API + types + renderResources + renderResourcesTemplate + splitMultiDoc + renderSingle)
    inputs.go    (buildInputSets + collectProviderInputs + decodeInputSet + MatchSelector + providerInputs)
    template.go  (slug init + execute + toYaml + mustToYaml — text/template plumbing)
    doc.go       (dedupKey + defaultNamespace + isClusterScoped + applyCommonMetadata + mergeStringMap + disabledByReconcileAnnotation)
    permute.go   (unchanged)

Net: render.go drops 523 → 229 LOC, focused on the public Render
entry point and the per-doc rendering pipeline. No behavior change;
identical imports re-distributed per file.

Verified:
- go test -count=1 ./...      — full suite green
- golangci-lint run ./...     — 0 issues
- tholinka/home-ops           — 118/0
- JJGadgets/Biohazard         — 199/0
- m00nwtchr/homelab-cluster   — 287/2 (pre-existing 404)
- buroa/k8s-gitops            — 73/0
- Zariel/home-ops             — 93/0
