+++ pkg/resourceset/render.go
@@ -31,20 +31,10 @@ package resourceset
 
 import (
 	"bytes"
-	"cmp"
-	"encoding/json"
 	"fmt"
-	"log/slog"
-	"slices"
-	"strings"
-	"text/template"
 
 	fluxopv1 "github.com/controlplaneio-fluxcd/flux-operator/api/v1"
-	sprig "github.com/go-task/slim-sprig/v3"
-	"github.com/gosimple/slug"
 	apix "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	"sigs.k8s.io/yaml"
 
 	"github.com/home-operations/flate/pkg/manifest"
@@ -126,160 +116,6 @@ func Render(rs *manifest.ResourceSet, resolve ProviderResolver) ([]map[string]an
 	return docs, nil
 }
 
-// providerInputs is one per-provider list of input sets carrying its
-// raw (un-normalized) name. Used to dispatch on spec.inputStrategy:
-// Flatten concatenates; Permute scopes by name and Cartesian-products.
-type providerInputs struct {
-	name   string             // raw provider name (rs.Name for inline; RSIP.Name otherwise)
-	inputs []map[string]any   // each set already carries its `provider` block
-}
-
-// buildInputSets gathers per-provider input lists then dispatches on
-// spec.inputStrategy. The ResourceSet's own inline spec.inputs are
-// treated as "provider 0", followed by each referenced
-// ResourceSetInputProvider in upstream's sorted order (matching
-// flux-operator/internal/inputs/combine.go).
-//
-// Flatten (default) concatenates: each input set is rendered once
-// against the templates, top-level access (`inputs.foo`).
-//
-// Permute Cartesian-products across providers: each result map has
-// the providers' input sets nested under their normalized names
-// (`inputs.<provider>.foo`) plus an `id` field — an adler32 hash of
-// the provider=index path matching upstream's flux-operator/internal/
-// inputs/permuter.go. Per-provider input lists are scoped before the
-// product so authors can disambiguate "rset.x" from "rsip.x". A 10000
-// permutation cap (also matching upstream) prevents pathological
-// Cartesian blowups.
-//
-// inputs.id is left to the provider when it's a Static RSIP; inline-
-// only inputs see no synthetic id under Flatten.
-func buildInputSets(rs *manifest.ResourceSet, resolve ProviderResolver) ([]map[string]any, error) {
-	groups, err := collectProviderInputs(rs, resolve)
-	if err != nil {
-		return nil, err
-	}
-
-	if rs.InputStrategy != nil && rs.InputStrategy.Name == fluxopv1.InputStrategyPermute {
-		return permute(groups, rs.InputStrategy.IncludeEmptyProviders)
-	}
-	// Default: Flatten — concatenate every provider's input list.
-	var out []map[string]any
-	for _, g := range groups {
-		out = append(out, g.inputs...)
-	}
-	return out, nil
-}
-
-// collectProviderInputs assembles the per-provider input lists: rs's
-// inline inputs as provider 0 (named after the ResourceSet itself),
-// then every resolved ResourceSetInputProvider in (namespace, name)
-// order. Each input set is stamped with a `provider` block so
-// templates can recover which CR sourced the values.
-func collectProviderInputs(rs *manifest.ResourceSet, resolve ProviderResolver) ([]providerInputs, error) {
-	var groups []providerInputs
-
-	// Provider 0: the ResourceSet itself with its inline spec.inputs.
-	inlineProv := map[string]any{
-		"apiVersion": fluxopv1.GroupVersion.String(),
-		"kind":       manifest.KindResourceSet,
-		"name":       rs.Name,
-		"namespace":  rs.Namespace,
-	}
-	inline := make([]map[string]any, 0, len(rs.Inputs))
-	for _, in := range rs.Inputs {
-		decoded := decodeInputSet(in)
-		decoded["provider"] = inlineProv
-		inline = append(inline, decoded)
-	}
-	groups = append(groups, providerInputs{name: rs.Name, inputs: inline})
-
-	if resolve == nil || len(rs.InputsFrom) == 0 {
-		return groups, nil
-	}
-
-	seen := make(map[string]struct{})
-	var providers []*manifest.ResourceSetInputProvider
-	for _, ref := range rs.InputsFrom {
-		matches, err := resolve(ref, rs.Namespace)
-		if err != nil {
-			return nil, fmt.Errorf("inputsFrom %q: %w", ref.Name, err)
-		}
-		for _, p := range matches {
-			if p == nil {
-				continue
-			}
-			k := p.Namespace + "/" + p.Name
-			if _, dup := seen[k]; dup {
-				continue
-			}
-			seen[k] = struct{}{}
-			providers = append(providers, p)
-		}
-	}
-	// Sort providers by (namespace, name) for deterministic output,
-	// matching upstream's Combine routine ordering.
-	slices.SortFunc(providers, func(a, b *manifest.ResourceSetInputProvider) int {
-		return cmp.Or(cmp.Compare(a.Namespace, b.Namespace), cmp.Compare(a.Name, b.Name))
-	})
-	for _, p := range providers {
-		exported, err := p.ExportedInputs()
-		if err != nil {
-			return nil, fmt.Errorf("ResourceSetInputProvider %s: %w", p.NamespacedName(), err)
-		}
-		if exported == nil && p.Type != "" && p.Type != fluxopv1.InputProviderStatic {
-			slog.Warn("resourceset: dynamic input provider contributes no inputs offline",
-				"resourceSet", rs.NamespacedName(),
-				"provider", p.NamespacedName(),
-				"type", p.Type)
-		}
-		provBlock := map[string]any{
-			"apiVersion": fluxopv1.GroupVersion.String(),
-			"kind":       manifest.KindResourceSetInputProvider,
-			"name":       p.Name,
-			"namespace":  p.Namespace,
-		}
-		pInputs := make([]map[string]any, 0, len(exported))
-		for _, set := range exported {
-			set["provider"] = provBlock
-			pInputs = append(pInputs, set)
-		}
-		groups = append(groups, providerInputs{name: p.Name, inputs: pInputs})
-	}
-	return groups, nil
-}
-
-func decodeInputSet(in fluxopv1.ResourceSetInput) map[string]any {
-	decoded := map[string]any{}
-	for k, v := range in {
-		if v == nil {
-			decoded[k] = nil
-			continue
-		}
-		var raw any
-		if err := json.Unmarshal(v.Raw, &raw); err != nil {
-			// Malformed entries are skipped silently — the parser
-			// already accepted the document, and there's no good
-			// signaling channel beyond a controller log line.
-			continue
-		}
-		decoded[k] = raw
-	}
-	return decoded
-}
-
-// MatchSelector returns true when sel matches lbls. Helper for
-// ProviderResolver implementations that filter by InputProviderReference.Selector.
-func MatchSelector(sel *metav1.LabelSelector, lbls map[string]string) (bool, error) {
-	if sel == nil {
-		return true, nil
-	}
-	s, err := metav1.LabelSelectorAsSelector(sel)
-	if err != nil {
-		return false, err
-	}
-	return s.Matches(labels.Set(lbls)), nil
-}
 
 // renderResources templates a single spec.resources entry once per
 // input set (or once with nil when inputs is empty), returning the
@@ -381,143 +217,3 @@ func renderSingle(tmplStr string, inputSet map[string]any) (map[string]any, erro
 	return doc, nil
 }
 
-// init mirrors upstream flux-operator's slugify configuration
-// (internal/builder/resourceset.go init): 63-char max with smart-
-// truncate, matching Kubernetes label-value limits. Without this,
-// slugify on inputs ≥64 chars renders the full slug in flate but a
-// truncated 63-char prefix in cluster — different downstream resource
-// names for the same ResourceSet input.
-func init() {
-	slug.MaxLength = 63
-	slug.EnableSmartTruncate = true
-}
-
-func execute(tmplStr string, inputSet map[string]any) ([]byte, error) {
-	tmpl, err := template.New("resourceset").
-		Delims("<<", ">>").
-		Funcs(sprig.HermeticTxtFuncMap()).
-		Funcs(template.FuncMap{
-			"slugify":    slug.Make,
-			"toYaml":     toYaml,
-			"mustToYaml": mustToYaml,
-			"inputs":     func() any { return inputSet },
-		}).
-		Option("missingkey=error").
-		Parse(tmplStr)
-	if err != nil {
-		return nil, fmt.Errorf("parse template: %w", err)
-	}
-	var buf bytes.Buffer
-	if err := tmpl.Execute(&buf, nil); err != nil {
-		return nil, fmt.Errorf("execute template: %w", err)
-	}
-	return buf.Bytes(), nil
-}
-
-// toYaml mirrors upstream's silent-error variant — encodes v as YAML;
-// any marshal error collapses to "". Templates expect this signature
-// so authors don't have to wrap every call in {{ if }}…{{ end }}.
-func toYaml(v any) string {
-	s, err := mustToYaml(v)
-	if err != nil {
-		return ""
-	}
-	return s
-}
-
-// mustToYaml is the explicit error-returning variant — surfaces a
-// marshal error to the template engine, which aborts execution. Use
-// this when the template is willing to fail loudly on malformed inputs.
-func mustToYaml(v any) (string, error) {
-	out, err := yaml.Marshal(v)
-	if err != nil {
-		return "", err
-	}
-	return strings.TrimRight(string(out), "\n"), nil
-}
-
-func dedupKey(doc map[string]any) string {
-	apiVersion, _ := doc["apiVersion"].(string)
-	kind, _ := doc["kind"].(string)
-	md, _ := doc["metadata"].(map[string]any)
-	name, _ := md["name"].(string)
-	ns, _ := md["namespace"].(string)
-	if kind == "" || name == "" {
-		return ""
-	}
-	return apiVersion + "|" + kind + "|" + ns + "|" + name
-}
-
-func defaultNamespace(doc map[string]any, ns string) {
-	if ns == "" {
-		return
-	}
-	md, _ := doc["metadata"].(map[string]any)
-	if md == nil {
-		md = map[string]any{}
-		doc["metadata"] = md
-	}
-	if cur, _ := md["namespace"].(string); cur != "" {
-		return
-	}
-	// Don't inject a namespace on cluster-scoped kinds — match the
-	// upstream operator behavior of leaving Namespace, ClusterRole etc.
-	// without a metadata.namespace.
-	if isClusterScoped(doc) {
-		return
-	}
-	md["namespace"] = ns
-}
-
-func isClusterScoped(doc map[string]any) bool {
-	kind, _ := doc["kind"].(string)
-	switch kind {
-	case "Namespace",
-		"ClusterRole", "ClusterRoleBinding",
-		"CustomResourceDefinition",
-		"PersistentVolume",
-		"StorageClass",
-		"PriorityClass",
-		"MutatingWebhookConfiguration", "ValidatingWebhookConfiguration",
-		"APIService",
-		"Node":
-		return true
-	}
-	return false
-}
-
-func applyCommonMetadata(doc map[string]any, cm *fluxopv1.CommonMetadata) {
-	if cm == nil || (len(cm.Labels) == 0 && len(cm.Annotations) == 0) {
-		return
-	}
-	md, _ := doc["metadata"].(map[string]any)
-	if md == nil {
-		md = map[string]any{}
-		doc["metadata"] = md
-	}
-	mergeStringMap(md, "labels", cm.Labels)
-	mergeStringMap(md, "annotations", cm.Annotations)
-}
-
-func mergeStringMap(md map[string]any, key string, in map[string]string) {
-	if len(in) == 0 {
-		return
-	}
-	out, _ := md[key].(map[string]any)
-	if out == nil {
-		out = make(map[string]any, len(in))
-	}
-	for k, v := range in {
-		out[k] = v
-	}
-	md[key] = out
-}
-
-func disabledByReconcileAnnotation(doc map[string]any) bool {
-	md, _ := doc["metadata"].(map[string]any)
-	ann, _ := md["annotations"].(map[string]any)
-	if v, _ := ann[fluxopv1.ReconcileAnnotation].(string); v == fluxopv1.DisabledValue {
-		return true
-	}
-	return false
-}
